Native data and metadata encryption for zfs
    Native Encryption for ZFS on OS X

    This change incorporates three major pieces:

    The first change is a keystore that manages wrapping
    and encryption keys for encrypted datasets. These
    commands mostly involve manipulating the new
    DSL Crypto Key ZAP Objects that live in the MOS. Each
    encrypted dataset has its own DSL Crypto Key that is
    protected with a user's key. This level of indirection
    allows users to change their keys without re-encrypting
    their entire datasets. The change implements the new
    subcommands "zfs load-key", "zfs unload-key" and
    "zfs change-key" which allow the user to manage their
    encryption keys and settings. In addition, several new
    flags and properties have been added to allow dataset
    creation and to make mounting and unmounting more
    convenient.

    The second piece of this patch provides the ability to
    encrypt, decyrpt, and authenticate protected datasets.
    Each object set maintains a Merkel tree of Message
    Authentication Codes that protect the lower layers,
    similarly to how checksums are maintained. This part
    impacts the zio layer, which handles the actual
    encryption and generation of MACs, as well as the ARC
    and DMU, which need to be able to handle encrypted
    buffers and protected data.

    The last addition is the ability to do raw, encrypted
    sends and receives. The idea here is to send raw
    encrypted and compressed data and receive it exactly
    as is on a backup system. This means that the dataset
    on the receiving system is protected using the same
    user key that is in use on the sending side. By doing
    so, datasets can be efficiently backed up to an
    untrusted system without fear of data being
    compromised.

    Reviewed by: Matthew Ahrens <mahrens@delphix.com>

+++ ZFSin/zfs/module/icp/include/sys/crypto/sched_impl.h
@@ -225,7 +225,7 @@ typedef	struct kcf_call_req {
 	kcf_req_params_t	kr_params;	/* Params saved for next call */
 	kcf_areq_node_t		*kr_areq;	/* Use this areq */
 	off_t			kr_saveoffset;
-	size_t			kr_savelen;
+	uint32_t			kr_savelen;
 } kcf_dual_req_t;
 
 /*
@@ -354,7 +354,7 @@ typedef struct kcf_context {
 typedef	struct kcf_ctx_template {
 	crypto_kcf_provider_handle_t	ct_prov_handle;	/* provider handle */
 	uint_t				ct_generation;	/* generation # */
-	size_t				ct_size;	/* for freeing */
+	uint32_t				ct_size;	/* for freeing */
 	crypto_spi_ctx_template_t	ct_prov_tmpl;	/* context template */
 							/* from the SW prov */
 } kcf_ctx_template_t;
@@ -490,11 +490,11 @@ extern kcf_prov_tried_t *kcf_insert_triedlist(kcf_prov_tried_t **,
     kcf_provider_desc_t *, int);
 extern kcf_provider_desc_t *kcf_get_mech_provider(crypto_mech_type_t,
     kcf_mech_entry_t **, int *, kcf_prov_tried_t *, crypto_func_group_t,
-    boolean_t, size_t);
+    boolean_t, uint32_t);
 extern kcf_provider_desc_t *kcf_get_dual_provider(crypto_mechanism_t *,
     crypto_mechanism_t *, kcf_mech_entry_t **, crypto_mech_type_t *,
     crypto_mech_type_t *, int *, kcf_prov_tried_t *,
-    crypto_func_group_t, crypto_func_group_t, boolean_t, size_t);
+    crypto_func_group_t, crypto_func_group_t, boolean_t, uint32_t);
 extern crypto_ctx_t *kcf_new_ctx(crypto_call_req_t  *, kcf_provider_desc_t *,
     crypto_session_id_t);
 extern int kcf_submit_request(kcf_provider_desc_t *, crypto_ctx_t *,
